+++ CHANGELOG.md
@@ -18,8 +18,10 @@ Before v3.0.0  |  v3.0.0
 - **Change:** Update `MiniApp.getCustomPermissions` & `MiniApp.setCustomPermissions` to return and store only the permissions listed in the Mini App's manifest.
 - **Fix:** Load ad error when do re-try loading.
 - **Fix:** Failure when simultaneous custom permission requests are received.
+- **Fix:** Exception for miniapp verification failed. See [this](miniapp/USERGUIDE.md#troubleshooting--faqs).
 - **Feature:** Added `MiniApp.getMiniAppManifest` interface to retrieve the manifest of a MiniApp.
 - **Feature:** Added `MiniApp.getCurrentManifest` interface to retrieve currently downloaded manifest of a MiniApp.
+- **Feature:** Added `MiniApp.getMiniAppManifest` interface to retrieve the meta-data of a MiniApp.
 
 **Sample App**
 - **Feature:** Added first-time launching screen to show the manifest information before downloading/launching a MiniApp.